Refactor to use [BottomSheetProductType] for `ManualProductTypeOptions`.
This also moves BottomSheetProductType into its own file (as it's now shared in different places). Localized strings are also updated.
